+++ ProjectSourceCode/Dice/Dice.h
@@ -22,7 +22,7 @@
 #include <regex>
 #include <string>
 
-#include "Observable.h"
+#include "..\Observable.h"
 
 using namespace observable;
